requirements.txt
opencv-python==4.12.0.88
torchnet==0.0.4
tqdm==4.67.1
scikit-image==0.21.0
pyqt5==5.15.11